Add snippet to README; Fix clang warnings in bundle_adjuster_keyframes* .

+++ include/keyframe_bundle_adjustment/internal/feature_point.hpp
@@ -13,17 +13,17 @@ struct FeaturePoint {
         ;
     }
 
-    FeaturePoint(float u, float v) : u(u), v(v), d(-1) {
+    FeaturePoint(double u, double v) : u(u), v(v), d(-1) {
         ;
     }
 
-    FeaturePoint(float u, float v, float d) : u(u), v(v), d(d) {
+    FeaturePoint(double u, double v, double d) : u(u), v(v), d(d) {
             ;
         }
 
-    float u;
-    float v;
-    float d;
+    double u;
+    double v;
+    double d;
 
     Eigen::Vector2d toEigen2d() const {
         return Eigen::Vector2d(u, v);
